{"types":"C","runs":4}

+++ Shualedruri2021.java
@@ -6,9 +6,9 @@ import acm.program.GraphicsProgram;
 
 public class Shualedruri2021 extends ConsoleProgram {
 	public void run() {
-	int a = 12;
+	double a = 12;
 	double b = 2.4;
-	int c = 2;
+	double c = 2.0;
 	double k = a/5;
 	println(k);
 }
